[TrustyAIService] Add fairness tests using DB storage
