--- tests/model_explainability/conftest.py
@@ -1,141 +1,0 @@
-from typing import Generator, Any
-
-import pytest
-import yaml
-from kubernetes.dynamic import DynamicClient
-from ocp_resources.config_map import ConfigMap
-from ocp_resources.deployment import Deployment
-from ocp_resources.namespace import Namespace
-from ocp_resources.pod import Pod
-from ocp_resources.secret import Secret
-from ocp_resources.service import Service
-from ocp_resources.service_account import ServiceAccount
-from ocp_resources.trustyai_service import TrustyAIService
-
-from tests.model_explainability.constants import TRUSTYAI_SERVICE
-from utilities.constants import MODELMESH_SERVING
-from utilities.infra import update_configmap_data
-
-MINIO: str = "minio"
-OPENDATAHUB_IO: str = "opendatahub.io"
-
-
-@pytest.fixture(scope="class")
-def trustyai_service_with_pvc_storage(
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    modelmesh_serviceaccount: ServiceAccount,
-    cluster_monitoring_config: ConfigMap,
-    user_workload_monitoring_config: ConfigMap,
-) -> Generator[TrustyAIService, Any, Any]:
-    with TrustyAIService(
-        client=admin_client,
-        name=TRUSTYAI_SERVICE,
-        namespace=model_namespace.name,
-        storage={"format": "PVC", "folder": "/inputs", "size": "1Gi"},
-        data={"filename": "data.csv", "format": "CSV"},
-        metrics={"schedule": "5s"},
-    ) as trustyai_service:
-        trustyai_deployment = Deployment(namespace=model_namespace.name, name=TRUSTYAI_SERVICE, wait_for_resource=True)
-        trustyai_deployment.wait_for_replicas()
-        yield trustyai_service
-
-
-@pytest.fixture(scope="class")
-def modelmesh_serviceaccount(
-    admin_client: DynamicClient, model_namespace: Namespace
-) -> Generator[ServiceAccount, Any, Any]:
-    with ServiceAccount(client=admin_client, name=f"{MODELMESH_SERVING}-sa", namespace=model_namespace.name) as sa:
-        yield sa
-
-
-@pytest.fixture(scope="session")
-def user_workload_monitoring_config(admin_client: DynamicClient) -> Generator[ConfigMap, Any, Any]:
-    data = {"config.yaml": yaml.dump({"prometheus": {"logLevel": "debug", "retention": "15d"}})}
-    with update_configmap_data(
-        client=admin_client,
-        name="user-workload-monitoring-config",
-        namespace="openshift-user-workload-monitoring",
-        data=data,
-    ) as cm:
-        yield cm
-
-
-@pytest.fixture(scope="class")
-def minio_pod(admin_client: DynamicClient, model_namespace: Namespace) -> Generator[Pod, Any, Any]:
-    with Pod(
-        client=admin_client,
-        name=MINIO,
-        namespace=model_namespace.name,
-        containers=[
-            {
-                "args": [
-                    "server",
-                    "/data1",
-                ],
-                "env": [
-                    {
-                        "name": "MINIO_ACCESS_KEY",
-                        "value": "THEACCESSKEY",
-                    },
-                    {
-                        "name": "MINIO_SECRET_KEY",
-                        "value": "THESECRETKEY",
-                    },
-                ],
-                "image": "quay.io/rh-ee-mmisiura/modelmesh-minio-examples:latest",
-                "name": MINIO,
-            }
-        ],
-        label={"app": "minio", "maistra.io/expose-route": "true"},
-        annotations={"sidecar.istio.io/inject": "true"},
-    ) as minio_pod:
-        yield minio_pod
-
-
-@pytest.fixture(scope="class")
-def minio_service(admin_client: DynamicClient, model_namespace: Namespace) -> Generator[Service, Any, Any]:
-    with Service(
-        client=admin_client,
-        name=MINIO,
-        namespace=model_namespace.name,
-        ports=[
-            {
-                "name": "minio-client-port",
-                "port": 9000,
-                "protocol": "TCP",
-                "targetPort": 9000,
-            }
-        ],
-        selector={
-            "app": MINIO,
-        },
-    ) as minio_service:
-        yield minio_service
-
-
-@pytest.fixture(scope="class")
-def minio_data_connection(
-    admin_client: DynamicClient, model_namespace: Namespace, minio_pod: Pod, minio_service: Service
-) -> Generator[Secret, Any, Any]:
-    with Secret(
-        client=admin_client,
-        name="aws-connection-minio-data-connection",
-        namespace=model_namespace.name,
-        data_dict={
-            "AWS_ACCESS_KEY_ID": "VEhFQUNDRVNTS0VZ",
-            "AWS_DEFAULT_REGION": "dXMtc291dGg=",
-            "AWS_S3_BUCKET": "bW9kZWxtZXNoLWV4YW1wbGUtbW9kZWxz",
-            "AWS_S3_ENDPOINT": "aHR0cDovL21pbmlvOjkwMDA=",
-            "AWS_SECRET_ACCESS_KEY": "VEhFU0VDUkVUS0VZ",  # pragma: allowlist secret
-        },
-        label={
-            f"{OPENDATAHUB_IO}/dashboard": "true",
-            f"{OPENDATAHUB_IO}/managed": "true",
-        },
-        annotations={
-            f"{OPENDATAHUB_IO}/connection-type": "s3",
-            "openshift.io/display-name": "Minio Data Connection",
-        },
-    ) as minio_secret:
-        yield minio_secret
