--- tests/model_explainability/constants.py
@@ -1,1 +1,0 @@
-TRUSTYAI_SERVICE: str = "trustyai-service"
